+++ custom_components/ldata/__init__.py
@@ -41,6 +41,7 @@ PLATFORMS: list[Platform] = [
     Platform.SELECT,
     Platform.SENSOR,
     Platform.SWITCH,
+    Platform.UPDATE,
 ]
 _LOGGER = logging.getLogger(LOGGER_NAME)